Move OIDC authentication commands to `cli/oidc.py`
