+++ commodore/cli/__init__.py
@@ -7,7 +7,6 @@ import click
 from dotenv import load_dotenv, find_dotenv
 from commodore import __git_version__, __version__
 from commodore.config import Config
-from commodore.login import login, fetch_token
 
 import commodore.cli.options as options
 
@@ -15,6 +14,7 @@ from .catalog import catalog_group
 from .component import component_group
 from .inventory import inventory_group
 from .package import package_group
+from .oidc import commodore_fetch_token, commodore_login
 
 
 def _version():
@@ -50,62 +50,8 @@ commodore.add_command(catalog_group)
 commodore.add_command(component_group)
 commodore.add_command(inventory_group)
 commodore.add_command(package_group)
-
-
-@commodore.command(
-    name="login",
-    short_help="Login to Lieutenant",
-)
-@options.api_url
-@options.oidc_discovery_url
-@options.oidc_client
-@options.pass_config
-def commodore_login(
-    config: Config, oidc_discovery_url: str, oidc_client: str, api_url: str
-):
-    """Login to Lieutenant"""
-    config.oidc_client = oidc_client
-    config.oidc_discovery_url = oidc_discovery_url
-    config.api_url = api_url
-
-    login(config)
-
-
-@commodore.command(
-    name="fetch-token",
-    short_help="Fetch Lieutenant token",
-)
-@options.api_url
-@options.oidc_discovery_url
-@options.oidc_client
-@options.pass_config
-@options.verbosity
-def commodore_fetch_token(
-    config: Config,
-    oidc_discovery_url: str,
-    oidc_client: str,
-    api_url: str,
-    verbose: int,
-):
-    """Fetch Lieutenant token
-
-    Get the token from the cache, if it's still valid, otherwise fetch a new token based
-    on the provided OIDC config and Lieutenant URL.
-
-    The command prints the token to stdout.
-    """
-    if api_url is None:
-        raise click.ClickException(
-            "Can't fetch Lieutenant token. Please provide the Lieutenant API URL."
-        )
-
-    config.api_url = api_url
-    config.oidc_client = oidc_client
-    config.oidc_discovery_url = oidc_discovery_url
-    config.update_verbosity(verbose)
-
-    token = fetch_token(config)
-    click.echo(token)
+commodore.add_command(commodore_login)
+commodore.add_command(commodore_fetch_token)
 
 
 def main():
